+++ R/kml_read.R
@@ -8,6 +8,10 @@
 #' If the source is a local file with a name ending in \code{.kmz}, 
 #' \code{kml_read} will treat it as a zipped KML file and will try to read its
 #' first file, as listed by \link[utils:unzip]{unzip}.
+#' 
+#' If the source is a link to Google My Maps, \code{kml_read} will try to 
+#' download the most recent version of the map and will then treat it as a 
+#' \code{.kmz} file.
 #' @param ... Arguments passed to \link[xml2:read_xml]{read_xml}, such as 
 #' \code{encoding} or \code{base_url}.
 #' See \link[xml2:read_xml]{read_xml} for details.
@@ -19,17 +23,37 @@
 #' kml_read(f)
 #' @seealso Google Developers. KML Reference.
 #' \url{https://developers.google.com/kml/documentation/kmlreference}
-#' @importFrom stringr %>% str_detect
+#' @importFrom stringr %>% str_detect str_extract
 #' @importFrom utils unzip
 #' @importFrom xml2 read_xml xml_ns
 #' @export
 kml_read <- function(x, ...) {
   
-  # case: local KMZ file
-  if (is.character(x) && file.exists(x) && str_detect(x, "\\.kmz$")) {
+  if (is.character(x) &&
+      str_detect(x, "^https?://(www\\.)?google\\.com/maps/d/(embed|viewer)")) {
+    
+    # case: live KML file from Google My Maps
+    
+    f <- tempfile(fileext = ".kmz")
+    x <- str_c("https://google.com/maps/d/kml?",
+               str_extract(x, "mid=(_|-|\\w)+"))
+    
+    message("KMZ: ", x)
+    
+    httr::GET(x) %>%
+      httr::content("raw") %>%
+      writeBin(f)
+
+    x <- kml_read(f)
+    unlink(f) # be clean
+    
+    return(x)
+    
+  } else if (is.character(x) && file.exists(x) && str_detect(x, "\\.kmz$")) {
+    # case: local KMZ file
   
-    y <- unzip(x, list = TRUE)[ 1 ]
-    message("File: ", y)
+    y <- unzip(x, list = TRUE)[1, 1]
+    message("KML: ", y)
     
     x <- read_xml(unz(x, y), ...)